+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # REPLACE-GLOBAL-WITH-NEXT-VERSION
 - MIG-3 - Fixes a problem with migrating order documents
+- MIG-5 - Improve snippet loading of DataSets
 - MIG-22 - Fixes a problem with migrating orders, caused by aborted orders
 - MIG-23 - Fixes a problem with log file download